Fixed typehints for fast blitter (for real this time I hope)
